+++ crates/store/src/db/schema.rs
@@ -47,6 +47,7 @@ diesel::table! {
     block_headers (block_num) {
         block_num -> BigInt,
         block_header -> Binary,
+        signature -> Binary,
     }
 }